Update date handling
I've split dateService.createFromDayMonthYear into createLocalTimeFromDayMonthYear and createUTCFromDayMonthYear.

The service manager is creating dates that refer to wall-clock time (the time on the wall) so these times need to be TZ aware.
The pupil date of birth needs to be in UTC as we are only storing the date component, so this ensures the date-of-birth is accurate (and not an hour before in the preceding day)