Added workaround for autoplay limitations on iOS + other devices
Some browsers will prevent sound from playing on a page until the
user 'acted' on the page (which is a key press or mouse click) and
Angular doesn't play nicely with it's form submission and those browsers,
so the first speech after logging in will not be read out because of it.

This is a workaround that speaks an empty string on the first event of such type,
in order to remove the autoplay limitations afterwards.