Update unique index check(school, pin)
This unique index prevents pupils from having the same login pin at the same
school (if that happened we couldn't identify the check to give the  pupil).

This update means the index only needs to be maintained for checks that are NEW
We don't care about pin collisions when the check is in other states as you can't log in to them.