Add prepare-check version 2 - which accepts a batch of prepared check objects for insertion into the preparedCheck table
This will cut down on network round trips from the admin app to the queue.  It was not possible to batch checks for all pupils in one go, as we
reached the queue message max message size of 64Kb.   Max batch size is hardcoded to 20.  Multiple network requests will be made if
needed.

This commit promisify's the queue service so we can detect any errors before sending the front-end user to a new page.