Feature/27594 send all pupils as one preparecheck message (#950)
* Refactor `prepare-check` to allow a new message version

* Add prepare-check version 2 - which accepts a batch of prepared check objects for insertion into the preparedCheck table

This will cut down on network round trips from the admin app to the queue.  It was not possible to batch checks for all pupils in one go, as we
reached the queue message max message size of 64Kb.   Max batch size is hardcoded to 20.  Multiple network requests will be made if
needed.

This commit promisify's the queue service so we can detect any errors before sending the front-end user to a new page.

* refactor to split out common `prepareEntity` code

* Factor out duplicate code

* lint fix

* lint fix

* lint fixes

* Refactor pupil status function in preparation for a new version 2 message

* Fix data in pupil register so it only shows live checks

* Add batch message capability to pupil-status

* used when creating checks to reduce network round trip latency

* lint fixes

* remove debug

* update test

* update test