Allow the app to start without the database
If the initial probe to the database fails, then the app will wait for 30 seconds and then exit.  This will be picked up by the PM2 node
process manager which will restart the app.  The wait time allows the db to come up, and keeps control even when the db is unavailable.