Feature/30640 admin app cannot recover from failed connection issues (#1059)
* Reduce number of default connections

* Avoid deprecation warning from hsts

* Bump mssql major version to v5

* Choose saner defaults for the maximum number of sql connections allowed

* Ensure the `isAuthenticated` flag is initialised

This prevents erros during template parsing

* Add a small test app to spike the behaviour of mssql under load

* Ensure we have a pool - we won't have it until the promise is resolved

* Add retry handling to the sqlService

* Remove blank line

* Allow the app to start without the database

If the initial probe to the database fails, then the app will wait for 30 seconds and then exit.  This will be picked up by the PM2 node
process manager which will restart the app.  The wait time allows the db to come up, and keeps control even when the db is unavailable.

* Lint fixes

* Add try / catch handlers

* Ensure connection timeout for the migrator is numeric

* Remove old config

* Make the wait time before exit configurable 'WAIT_TIME_BEFORE_EXIT' in seconds

Tidy up and sort the config file

* Update comment

* Fix typo