feature/31203 - retry logic for all sql consumers (#1086)
* update less-tedious to v2

* bump less-tedious to get back sqlService.initialise

* resync lock file

* update config assignment

* Fix unit tests inadvertantly attempting a db connection due to a side-effect

* Bump versions

* Add retry logic from less-tedious to admin app

* Add info on disabling functions

* Set maxPollingInterval to production setting - 60s

This is the default in Azure

From Azure documentation

"The queue trigger implements a random exponential back-off algorithm to reduce the effect of idle-queue polling on storage transaction costs. When a message is found, the runtime waits two seconds and then checks for another message; when no message is found, it waits about four seconds before trying again. After subsequent failed attempts to get a queue message, the wait time continues to increase until it reaches the maximum wait time, which defaults to one minute. The maximum wait time is configurable via the maxPollingInterval property in the host.json file."

* Set visibility timeout to production setting: 30s

`visibilityTimeout` is "The time interval between retries when processing of a message fails."

* Set `batchSize` to production setting: 8

The Azure default batchsize is 16.  We are setting it lower as a bit of guess, but we have seen database connention
when running at a batchSize of 16 at high load.

* Set `maxDequeueCount` to productions settings: 3

The Azure default is 5 - but this seems overkill.  Messages on the poison queue are usually un-processable.

* Remove version 1 stanza

* Bug: change 60s to 1 min

* Add sqlService to lib

* Replace less-tedioous with local sqlService

* Update paths

* Consolidate to a single pool

* Remove outdated use of less-tedious

* Remove outdated use of winston

* Remove winston and less-tedious from functions

* Remove lint

* Bump versions

* Add config for the db retry mechanism

New Env Vars:

RETRY_MAX_ATTEMPTS
RETRY_PAUSE_MS
RETRY_PAUSE_MULTIPLIER

* Add tunable env vars to the functions for DB retry

* upped timeout for feedback scenario

* remove sql config duplication

* remove unused var

* enhanced debug logging and config.

* fix docker build issue and reset function behaviour settings to see if that fixes travis problem (as new host.json settings mean functions wait 30 secs on startup before picking anything up

* undo function polling behaviour changes

* increasing wait and printing var's

* Remove instances of less-tedious

* Restart the functions container so it can connect to the db

* Set the SQL ENCRYPT var in the functions Dockerfile

* Fix a failing unit test (an SEP)

* Apply sorting to the config

* Add comment

* Ensure that travis also stops and removes the mtc_funcs container

* Use `yarn` consistently