[admin] - fix query for selecting pupils not attending
We need to include pupils who have hhad a live check, but it has expired as long as they don't have any other live checks.
It seems easier to just say that the pupil's process status must be NOT STARTED if they want to set them to not attending.

The NOT STARTED process status is equivalent to the UNALLOC pupil status.