[admin-app] fix docker build
1. copy assets into an `assets` directory on docker
2. copy `sql.config.js` into a `config` directory, rather than a file called `config`
3. Add dependency `auth-modes.js`

I suspect that this has been broken forever.  The gulp upgrade was the trigger as it now bombs out
during `bundlejs` on non-existent files.

I really don't understand why we don't copy the whole app over, which would reduce the maintenance burden
on this Dockerfile

If the config is broken then the session-expiry file will be knackered which breaks the uglify process.

This is all pretty horrible.