[functions]  add flag to follow redirects
It looks like the dotnet install scipt has moved and it is now serving a redirect on this location.  Furthermore, the docs still suggest the current URL is the direct one, which it isn't.  Perhaps they will update soon.

The redirect is to: https://dotnet.microsoft.com/download/dotnet-core/scripts/v1/dotnet-install.sh