Bug/Map datetimeoffset to mssql datetimeoffset instead of datetime (#1394)
* Map datetimeoffset to mssql datetimeoffset instead of datetime

* Added datetime2 and datetime as types within sql service

* removed xit

* removed comments, console logs and changed test descriptions

* [functions]  add flag to follow redirects

It looks like the dotnet install scipt has moved and it is now serving a redirect on this location.  Furthermore, the docs still suggest the current URL is the direct one, which it isn't.  Perhaps they will update soon.

The redirect is to: https://dotnet.microsoft.com/download/dotnet-core/scripts/v1/dotnet-install.sh

* [functions] update Dockerfile

* [functions] revert Dockfile change