[admin] update pupils eligible for restart view
This fixes a bug caused by the new introduction of the new status 'VOID'.  Once a check is restarted it is set to VOID as that check should no longer be used.  The new status was not recognised by the view that shows eligible restarts, which meant the answer was always 1 (it only detected the check that was itself being restarted), and so a restart was allowed.
To fix this I have altered the restart eligibility view so that the check counts towards the pupils restart count once they log in.  This is because the "Started" status will be removed very soon.