[admin] support long URLs
Adding 200 pupils results in a 9000 byte URL mostly made up of UUIDs to highlight on the page shown to the user after adding the pupils.
Whilst node used to have an 80KB header limit this was changed in 10.14.0 [1]  to address a high-severity vulnerability[2] This reduced the
max length of the HTTP headers to 8192 bytes.

This commit inceases the max length of the headers to 15000 bytes which is sufficient for 200 pupils.

Please note that the issue is mitigated by the use of a proxy.

[1] https://github.com/nodejs/node/blob/v10.x/doc/changelogs/CHANGELOG_V10.md
[1]  https://nodejs.org/en/blog/vulnerability/november-2018-security-releases/#denial-of-service-with-large-http-headers-cve-2018-12121