[tslib] update comment
After investigation it appears to be impossible to use the MS entityGenerator to actual type the markedAt date as
a date rather than the string I get in testing.

This works when used in an application, but when using the functionBindings for output it fails.

`azure-storage` is now legacy with no apparent upgrade option for TableStorage.