[admin] remove lint from the front-end tests
Karma - jquery is packaged in the app.js - so need to include it in the frameworks section

We are only testing on Chrome, so need to include extra launchers.  Although, we should clearly test on other browsers too.

check-forms.js and mtc-check-forms.js are dulicate functions stored under different global objects.   The preference for MTC functions is to store them under 'MTCAdmin' and leave 'GOVUK' for GDS stuff.