Feature/41849 spike upgrade jquery (#1675)
* [admin]  add jquery v3

* [admin] whitespace

* [admin] jquery v3 fix

* [admin] jquery v3 fix

* [admin] update comment

* [admin] remove old jQuery

* [admin]  don't lint vendor file

* [admin] remove lint from the front-end tests

Karma - jquery is packaged in the app.js - so need to include it in the frameworks section

We are only testing on Chrome, so need to include extra launchers.  Although, we should clearly test on other browsers too.

check-forms.js and mtc-check-forms.js are dulicate functions stored under different global objects.   The preference for MTC functions is to store them under 'MTCAdmin' and leave 'GOVUK' for GDS stuff.

* [admin] FE: remove duplicate code and lint

Co-authored-by: Mohsen Qureshi <activemq@users.noreply.github.com>