Feature/42055 ctf file download (#1678)
* [admin] add new route for ctf-download

* [admin] Add ctfService

Just validation so far

* [admin] add new error type

Set status to 404 if the client requests the CTF xml file before the business rules allow

* [admin] add attendanceCode to the results dataset

* [admin] add transformation

* [admin] add tests

* [admin] Update dependencies

Upgrade redis-connect to v4.  Now uses external client - ioredis client from cache service.

* [admin] fix jsdoc

* [admin] bug fix

Without this it errors out with a message that it can't find the error method

* [admin] update jsdoc

* [admin] add reviver for Redis Cache deserialisation

Revive ISO8601 strings to Moment objects

* [admin] add ctf download

* [admin] ctf refactor

By default the CTF XML file will be downloaded as an attachement.  But add `inline=1` to the QS to view it in the browser if you prefer.

* [admin] add doc

* [admin bug fix case

* [admin] add link to download the results from the results screen

* [admin] remove sql service bug

As redis-cache will now revive ISO 8601 dates this line is no longer needed, which is good, as it doesn't always work.

* [admin] fix test

* [admin] add test

* [admin] spelling

* [admin] return value by default

* [admin] return value by default

* [admin] js idiom, just in case

* cache bust update

* [admin] cache invalidation

invalidate school results cache on: changing a pupil's not attending status.  Don't forget the HDF method too.

* updated tests

Co-authored-by: Mohsen Qureshi <activemq@users.noreply.github.com>
Co-authored-by: Mohsen Qureshi <mjq@Mohsens-Mac-mini.local>