[admin] remove moment
Type errors are liable to arise between moment-timezone and moment as moment-timezone installs it's own version of moment.  It seems easier just to install that version,
which at the time of writing is at n-1 versions out of date, and n+5 versions ahead of our installed version.

Win-win: we don't have to add ts-ignore messages everywhere and we get a newer version of moment.