[tslib] add new function
Add `school-results-cache-determiner`

This function will determine the date/time is now between the close of the live check window and the opening of the results on the
monday after.  If so, it will send a message per school onto the `school-results-cache` queue so another function can cache it.

New config environment variable: SCHOOL_RESULTS_CACHE [0|1|2]

Default value is to cache if in date (production setting). For local development you probably want to add this line to your `.env` file:

// Always cache, irrespective of date
SCHOOL_RESULTS_CACHE=2