Feature/41979 cache school results in redis from Azure Functions (#1704)
* [infra] add 2 new service bus queues

* [infra] sort queue names

* [functions-app] modify sync-results-to-db function

* added output binding to send message on completion

Note: the message is sent whether any activity was performed or not.

* [admin]  set ttl for service bus queue to 1 day

Add service-bus node modules installation into start.sh

* Add new function app skel

* [tslib] add new function

Add `school-results-cache-determiner`

This function will determine the date/time is now between the close of the live check window and the opening of the results on the
monday after.  If so, it will send a message per school onto the `school-results-cache` queue so another function can cache it.

New config environment variable: SCHOOL_RESULTS_CACHE [0|1|2]

Default value is to cache if in date (production setting). For local development you probably want to add this line to your `.env` file:

// Always cache, irrespective of date
SCHOOL_RESULTS_CACHE=2

* Add documentation

* Update docs

* Update comment

* [tslib] add new function skel

* Add func name to log messages

* wip - refactor pupil identification flag service in ts

New approach as the old pattern had an outstanding bug which is now fixed.

* [tslib] finish refactor

* [tslib] port tableSorting service to ts

* bugfix: use the right config

* Prevent timer from running if it is past due

* Remove unused function

This function in functions-app is now replaced by the typescript one, running from a timer, in func-consumption `school-results-cache`

* Add debugging for ci server

* Fix race condition in unit tests

* Update edit_pupil_steps.rb

Co-authored-by: Guy Harwood <GuyHarwood@users.noreply.github.com>
Co-authored-by: Mohsen Qureshi <mjq@Mohsens-Mac-mini.local>