Feature/44783 ps report func 2 (#1783)
* [tslib] add new function skeleton

* [tslib] bug fix timer trigger definition

* [tslib] change line ending to \n

* [tslib] Initial data service work for retrieving pupil results

WIP

* [tslib] add check entity

* Add doc

* [tslib] add checkForm

* [tslib] add answers

* [tslib] add inputs

* [tslib] add device

* Whitespace changes only  - convert newlines to \n

* [tslib] add events

* [tslib] Type incoming message

* Tweaks

* Update doc

* Fix a test

Typo was causing the live check end date to not be setup correctly.  The default is 'now plus 4 days',
so this bug was only apparent inthe last 4 days of the year.

* add tests

* add tests

* cherry pick fix for test

* add documentation

* refactor

remove all duplicate calls to `getSchool()`

* remove TODO item

* refactor

cache the checkForm lookups as there are only a small number of checkForms

* whitespace change only

newlines converted to \n

* update function gulpfiles

refactor for gulp v4
remove duplicate step in `yarn build` to remove spec files from the dist folder.  Not needed as this is covered in the gulp task.

* [func-throttled] Add watch task to gulp

* Freeze outputs

* Mod gulpfiles

Only run watch when called by start:dev

* [tslib, func-consumption, func-throttled] upgrade deps

* Convert line endings to unix standard

* Clean up TODO items

remove a method
add documentation

* Remove dev packages that are no longer used

* Update docs