Bugfix for AA keypad removal
When the keypad was removed via an Access Arrangement the SPA was crashing during question initialisation
as the buttons it was binding the listener functions to was not present.

The fix detects the presence of the button before binding, rather than the corresponding AA as this is deemed to be more reliable.