Feature/46532 the retro input assistant link does not display until official check (#1875)
* [admin] bugfix: don't show the retro input assistent until the check window opens

* [admin] add tests

We want these tests to be unit tests rather than integration tests because the driver for showing it is the date and the checkwindow.  If we
were to make these  ntegration tests we can easily test the output, but if we have to mock the app then we would need to start and stop the app
for each test which would be slow.

This approach has the benefit that we can test the data presented to the view, without needing the view content itself.  The view here is entirely dumb
which seems best.

* [admin] fix test suite

Co-authored-by: Mohsen Qureshi <activemq@users.noreply.github.com>