Bug/41813 pupil identification flag service (#1983)
* [admin] refactor pupilIdentificationFlagService

* add new method `sortAndAddIdentificationFlagService`

The Pupil Identification Flag Service has a requirement that the incoming pupil arrary is sorted or it will not work.  The new method
does the sort and adds the disambiguation data, so is now preferred.

* [admin] refactor getPupilsEligibleForPinGeneration

* [admin] refactor pupil-presenter

* [admin] refactor pupils-not-taking-check

* [admin] refactor pupil status service

* [admin] refactor group service

* [admin] refactor pin generation v2 service

* [admin] refactor pupil generation v2 data service

* [admin] refactor pupil access arrangements service

* [admin] refactor pupil register v2 service

* [admin] remove commented out line

* [admin] refactor pupils not taking check

* [admin] refactor restart v2 service

* [admin] refactor restart v2 service

* [admin] refactor result service

* [admin] refactor and clean-up

* [admin] use new formatted version of the dob

Avoid using a prop that can be a moment OR string data types

* [admin] bugfix + tests

* [admin] spec change

In order to align with the QA tests `showDoB` will always be true when `showMiddlenames` is true.  This makes the functionality exactly the same as the previous code.

* update doc

* [admin] refactor restarts

Remove dependency on side-effects

* [admin] bug fix

Middlenames should only show if the names are the same, and the DoB are the same.